修正模板消息的返回值，返回msgId。 会员卡接口的调整。
