+++ weixin4j-mp/src/main/java/com/foxinmy/weixin4j/mp/WeixinProxy.java
@@ -1005,7 +1005,8 @@ public class WeixinProxy {
 	 *      删除群发</a>
 	 * @see com.foxinmy.weixin4j.mp.api.MassApi
 	 * @see {@link #massByGroupId(Tuple, int)}
-	 * @see {@link #massByOpenIds(Tuple, String...)
+	 * @see {@link #massByOpenIds(Tuple, String...)
+
 	 *
 	 */
 	public ApiResult deleteMassNews(String msgid) throws WeixinException {
@@ -1570,7 +1571,7 @@ public class WeixinProxy {
 	 * @seee com.foxinmy.weixin4j.msg.event.TemplatesendjobfinishMessage
 	 * @see com.foxinmy.weixin4j.mp.api.TmplApi
 	 */
-	public ApiResult sendTmplMessage(TemplateMessage tplMessage)
+	public String sendTmplMessage(TemplateMessage tplMessage)
 			throws WeixinException {
 		return tmplApi.sendTmplMessage(tplMessage);
 	}
