+++ weixin4j-mp/src/main/java/com/foxinmy/weixin4j/mp/api/CardApi.java
@@ -13,9 +13,14 @@ import com.foxinmy.weixin4j.model.Token;
 import com.foxinmy.weixin4j.model.card.CardCoupon;
 import com.foxinmy.weixin4j.model.card.CardCoupons;
 import com.foxinmy.weixin4j.model.card.CardQR;
+import com.foxinmy.weixin4j.model.card.MemberInitInfo;
+import com.foxinmy.weixin4j.model.card.MemberUpdateInfo;
+import com.foxinmy.weixin4j.model.card.MemberUserForm;
+import com.foxinmy.weixin4j.model.card.MemberUserInfo;
 import com.foxinmy.weixin4j.model.qr.QRParameter;
 import com.foxinmy.weixin4j.model.qr.QRResult;
 import com.foxinmy.weixin4j.token.TokenManager;
+import com.foxinmy.weixin4j.type.card.CardStatus;
 import com.foxinmy.weixin4j.type.card.CardType;
 import com.foxinmy.weixin4j.util.IOUtil;
 
@@ -191,7 +196,7 @@ public class CardApi extends MpApi {
 	 * @author fengyapeng
 	 * @since 2016 -12-20 11:48:23
 	 */
-	public void getCardStatus(String cardId) throws WeixinException {
+	public CardStatus queryCardStatus(String cardId) throws WeixinException {
 		JSONObject requestObj = new JSONObject();
 		requestObj.put("card_id",cardId);
 		String card_get_uri = getRequestUri("card_get_uri");
@@ -201,7 +206,8 @@ public class CardApi extends MpApi {
 		JSONObject card = responseAsJson.getJSONObject("card");
 		String cardType = card.getString("card_type");
 		JSONObject baseInfo = card.getJSONObject(cardType.toLowerCase()).getJSONObject("base_info");
-		baseInfo.get("status");
+		String status = baseInfo.getString("status");
+		return CardStatus.valueOf(status);
 	}
 
 	/**
@@ -226,6 +232,75 @@ public class CardApi extends MpApi {
 	    JSONObject jsonObject= 	response.getAsJson();
 	    return jsonObject.getBoolean("send_check");
 	}
+
+
+
+
+
+	/**
+	 * 激活方式说明
+	 * 接口激活通常需要开发者开发用户填写资料的网页。通常有两种激活流程：
+	 * 1. 用户必须在填写资料后才能领卡，领卡后开发者调用激活接口为用户激活会员卡；
+	 * 2. 是用户可以先领取会员卡，点击激活会员卡跳转至开发者设置的资料填写页面，填写完成后开发者调用激活接口为用户激活会员卡。
+	 *
+	 * @see <a href='https://mp.weixin.qq.com/wiki?action=doc&id=mp1451025283&t=0.8029895777585161#6.1'>接口激活</a>
+	 */
+	public ApiResult activate(MemberInitInfo memberInitInfo) throws WeixinException {
+		String card_member_card_activate_uri = getRequestUri("card_member_card_activate_uri");
+		Token token = tokenManager.getCache();
+		WeixinResponse response = weixinExecutor
+				.post(String.format(card_member_card_activate_uri, token.getAccessToken()), JSON.toJSONString(memberInitInfo));
+		return response.getAsResult();
+	}
+
+	/**
+	 * 设置开卡字段接口
+	 * 开发者在创建时填入wx_activate字段后，
+	 * 需要调用该接口设置用户激活时需要填写的选项，否则一键开卡设置不生效。
+	 *
+	 * @see <a href='https://mp.weixin.qq.com/wiki?action=doc&id=mp1451025283&t=0.8029895777585161#6.2'>一键激活</a>
+	 */
+	public ApiResult setActivateUserForm(MemberUserForm memberUserForm) throws WeixinException {
+		String user_form_uri = getRequestUri("card_member_card_activate_user_form_uri");
+		Token token = tokenManager.getCache();
+		WeixinResponse response = weixinExecutor
+				.post(String.format(user_form_uri, token.getAccessToken()), JSON.toJSONString(memberUserForm));
+		return response.getAsResult();
+	}
+
+	/**
+	 * 拉取会员信息接口。
+	 *
+	 * @param cardId the card id
+	 * @param code   the code
+	 * @author fengyapeng
+	 * @since 2016 -12-21 11:28:45
+	 */
+	public MemberUserInfo getMemberUserInfo(String cardId, String code) throws WeixinException {
+		String user_info_uri = getRequestUri("card_member_card_user_info_uri");
+		Token token = tokenManager.getCache();
+		JSONObject jsonObject = new JSONObject();
+		jsonObject.put("card_id", cardId);
+		jsonObject.put("code", code);
+		WeixinResponse response = weixinExecutor.post(String.format(user_info_uri, token.getAccessToken()), JSON.toJSONString(jsonObject));
+		return response.getAsObject(new TypeReference<MemberUserInfo>() {
+		});
+	}
+
+	/**
+	 * 更新会员
+	 * @param updateInfo
+	 * @return
+	 * @throws WeixinException
+	 */
+	public JSONObject updateUserInfo(MemberUpdateInfo updateInfo) throws WeixinException {
+		String card_member_card_update_user_uri = getRequestUri("card_member_card_update_user_uri");
+		Token token = tokenManager.getCache();
+		WeixinResponse response = weixinExecutor
+				.post(String.format(card_member_card_update_user_uri, token.getAccessToken()), JSON.toJSONString(updateInfo));
+		return response.getAsJson();
+	}
+
